# specs/005-contracts-module/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FlyAccounts
  description: 'FlyAccounts API 6.0.0: global Clients CRUD; server-side search/pagination on clients,
    contracts, people, roles; contracts require clientId on create.'
  version: 6.0.0
servers:
- url: /v1
  description: Versioned API prefix
...
paths:
...
  /clients:
    get:
      operationId: listClients
      summary: List clients with search and pagination
      description: Requires manage_contracts. Global client master (not entity-scoped).
      tags:
      - Clients
      security:
      - bearerAuth: []
      parameters:
      - *id001
      - *id002
      - *id003
      responses:
        '200':
          description: Paginated clients.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientsResponse'
        '401': &id004
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403': *id004
    post:
      operationId: createClient
      summary: Create a client
      description: Requires manage_contracts. Name must be unique (case-insensitive).
      tags:
      - Clients
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Client created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400': *id004
        '401': *id004
        '403': *id004
        '409':
          description: Duplicate client name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          example:
            code: duplicate_client_name
            message: A client with that name already exists.
  /clients/{clientId}:
    get:
      operationId: getClient
      summary: Get a client
      tags:
      - Clients
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Client.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '401': *id004
        '403': *id004
        '404': *id004
    patch:
      operationId: updateClient
      summary: Update a client
      tags:
      - Clients
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest'
      responses:
        '200':
          description: Updated client.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400': *id004
        '401': *id004
        '403': *id004
        '404': *id004
        '409': *id004
    delete:
      operationId: deleteClient
      summary: Delete a client
      description: Refused with client_in_use when non-deleted contracts reference the client.
      tags:
      - Clients
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ClientId'
      responses:
        '204':
          description: Deleted.
        '401': *id004
        '403': *id004
        '404': *id004
        '409':
          description: Client is referenced by contracts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          example:
            code: client_in_use
            message: This client is linked to 2 contracts.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: FlyAccounts access JWT from POST /auth/microsoft or POST /auth/refresh.
  parameters:
...
    ClientId:
      name: clientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - invalid_token
          - personal_account
          - unknown_tenant
          - unauthorized
          - forbidden
          - pending_access
          - not_found
          - duplicate_assignment
          - last_admin_required
          - duplicate_invite
          - already_present
          - duplicate_role_name
          - role_still_assigned
          - duplicate_permission
          - validation_error
          - entity_context_required
          - invalid_currency
          - invalid_file_type
          - client_in_use
          - duplicate_client_name
        message:
          type: string
...
    Client:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - address
      - contactPerson
      - contactEmail
      - contactPhone
      - vatNumber
      - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        contactPerson:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
        vatNumber:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    ClientsResponse:
      type: object
      additionalProperties: false
      required:
      - clients
      - page
      - pageSize
      - total
      properties:
        clients:
          type: array
          items:
            $ref: '#/components/schemas/Client'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          enum:
          - 10
          - 25
          - 50
        total:
          type: integer
          minimum: 0
    CreateClientRequest:
      type: object
      additionalProperties: false
      required:
      - name
      - address
      - contactPerson
      - contactEmail
      - contactPhone
      - vatNumber
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        contactPerson:
          type: string
          minLength: 1
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
          minLength: 1
        vatNumber:
          type: string
          minLength: 1
        notes:
          type: string
    UpdateClientRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        contactPerson:
          type: string
          minLength: 1
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
          minLength: 1
        vatNumber:
          type: string
          minLength: 1
        notes:
          type: string